cdk/requirements.txt
attrs==19.3.0
aws-cdk.assets==1.59.0
aws-cdk.aws-applicationautoscaling==1.59.0
aws-cdk.aws-autoscaling-common==1.59.0
aws-cdk.aws-certificatemanager==1.59.0
aws-cdk.aws-cloudfront==1.59.0
aws-cdk.aws-cloudwatch==1.59.0
aws-cdk.aws-codeguruprofiler==1.59.0
aws-cdk.aws-ec2==1.59.0
aws-cdk.aws-efs==1.59.0
aws-cdk.aws-events==1.59.0
aws-cdk.aws-iam==1.59.0
aws-cdk.aws-kms==1.59.0
aws-cdk.aws-lambda==1.59.0
aws-cdk.aws-logs==1.59.0
aws-cdk.aws-route53==1.59.0
aws-cdk.aws-s3==1.59.0
aws-cdk.aws-s3-assets==1.59.0
aws-cdk.aws-s3-deployment==1.59.0
aws-cdk.aws-sqs==1.59.0
aws-cdk.aws-ssm==1.59.0
aws-cdk.cloud-assembly-schema==1.59.0
aws-cdk.core==1.59.0
aws-cdk.cx-api==1.59.0
aws-cdk.region-info==1.59.0
cattrs==1.0.0
colorama==0.4.3
constructs==3.0.4
crayons==0.1.2
jsii==1.10.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.2
